requirements.txt
python-dotenv
streamlit
langchain
langchain_community
langchain_openai
langchain-text-splitters
libmagic
faiss-cpu